Remove Wikipedia scraping
The Wikipedia page has been removed. We _could_ get that data back from
pointing at the last version of it before the redirect was added, but
ideally all the people we need _should_ be found from the SPARQL query
anyway. If they're not currently being returned from that, we should add
relevant P39s.